Add multi-user management support.

+++ src/opwifi/app/Http/Controllers/Opwifi/OwCRUDController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\Opwifi;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Pagination\Paginator;
-use Input;
+use Input, Auth;
 
 use App\Http\Helpers\Opwifi\SqlFakeScheduler;
 
@@ -13,6 +13,13 @@ abstract class OwCRUDController extends Controller
 {
 	public function __construct() {
 		SqlFakeScheduler::update();
+	}
+
+	private function getLimitUserId() {
+		if (isset($this->limitUserId) &&
+				Auth::User()['right'] != 'admin')
+			return Auth::User()['id'];
+		return 0;
 	}
 
 	public function getSelect(Request $request) {
@@ -32,6 +39,11 @@ abstract class OwCRUDController extends Controller
 			});
 		}
 		$query = $this->newOwnModel();
+
+		$userId = $this->getLimitUserId();
+		if ($userId > 0)
+			$query = $query->where($this->limitUserId, $userId);
+
 		if ($tag_ids && isset($this->indexOwnModelTag)) {
 			$ids = array();
 			$rales = $this->newOwnModelTagRelationships()->whereIn('tag_id', explode(',',$tag_ids))->get();
@@ -56,12 +68,22 @@ abstract class OwCRUDController extends Controller
 		if (!$request->isJson()) {
 			return ;
 		}
+		$userId = $this->getLimitUserId();
 		$cfgs = $request->json()->all();
 		foreach ($cfgs as $cfg) {
-			if ($this->indexOwnModel && !$cfg[$this->indexOwnModel]) continue;
+			if ($this->indexOwnModel && !isset($cfg[$this->indexOwnModel])) continue;
 			if (isset($this->hasOwnModelDefault) && $this->hasOwnModelDefault) {
 				if ($this->newOwnModel()->where('id', 1)->count() == 0) {
 					$cfg['id'] = 1;
+				}
+			}
+			if ($userId > 0) {
+				$cfg[$this->limitUserId] = $userId;
+				/* Try get first, if empty in limitUserId, update it */
+				$old = $this->newOwnModel()->where($this->indexOwnModel, $cfg[$this->indexOwnModel])->first();
+				if ($old && $old[$this->limitUserId] == null) {
+					$old->update($cfg);
+					continue;
 				}
 			}
 			if (method_exists($this, 'createOwnModel')) {
@@ -78,12 +100,22 @@ abstract class OwCRUDController extends Controller
 		if (!$request->isJson() || !$root) {
 			return ;
 		}
+		$userId = $this->getLimitUserId();
 		$cfgs = $request->json()->all();
 		foreach ($cfgs as $cfg) {
-			if ($this->indexOwnModel && !$cfg[$this->indexOwnModel]) continue;
+			if ($this->indexOwnModel && !isset($cfg[$this->indexOwnModel])) continue;
 			if (isset($this->hasOwnModelDefault) && $this->hasOwnModelDefault) {
-				if ($this->newOwnModel()->where('id', 1)->count() == 0) {
+				if ($this->newOwnModelRoot()->where('id', 1)->count() == 0) {
 					$cfg['id'] = 1;
+				}
+			}
+			if ($userId > 0) {
+				$old = $this->newOwnModelRoot()->where($this->indexOwnModel, $cfg[$this->indexOwnModel])->first();
+				if ($old && $old[$this->limitUserId] == null) {
+					$own = $this->getOwnModelByRoot($old);
+					$old->update($cfg);
+					$own->update([$this->limitUserId => $userId]);
+					continue;
 				}
 			}
 			if (method_exists($this, 'createOwnModelRoot')) {
@@ -99,9 +131,12 @@ abstract class OwCRUDController extends Controller
 		if (!$request->isJson()) {
 			return ;
 		}
+		$userId = $this->getLimitUserId();
 		$cfgs = $request->json()->all();
 		foreach ($cfgs as $cfg) {
 			if (!$cfg['id']) continue;
+			if ($userId > 0 && $userId != $cfg[$this->limitUserId])
+				return response()->json(['success'=>false]);
 			$this->newOwnModel()->where('id', $cfg['id'])->delete();
 		}
 		return response()->json(['success'=>true]);
@@ -112,17 +147,23 @@ abstract class OwCRUDController extends Controller
 		if (!$request->isJson() || !$root) {
 			return ;
 		}
+		$userId = $this->getLimitUserId();
 		$cfgs = $request->json()->all();
 		foreach ($cfgs as $cfg) {
 			if (!$cfg['id']) continue;
 			$meta = $this->newOwnModel()->where('id', $cfg['id'])->first();
-			if ($meta)
+			if ($meta) {
+				if ($userId > 0 && $userId != $meta[$this->limitUserId])
+					return response()->json(['success'=>false]);
 				$meta->$root()->delete();
+			}
 		}
 		return response()->json(['success'=>true]);
 	}
 
 	public function postUpdate(Request $request) {
+		$userId = $this->getLimitUserId();
+
 		$id = $request->get('id');
 		if (!$id) {
 			$id = $request->get('pk');
@@ -156,6 +197,14 @@ abstract class OwCRUDController extends Controller
 				}
 			}
 			unset($cfg['id']);
+			if ($userId > 0 && (
+					(isset($cfg[$this->limitUserId]) &&
+					$userId != $cfg[$this->limitUserId]) ||
+					($it[$this->limitUserId] != null &&
+					$userId != $it[$this->limitUserId])
+					)) {
+				return response()->json(['success'=>false]);
+			}
 			if (count($cfg)) {
 				$it->update($cfg);
 			}
